Fixed a bug with the cyclomatic complexity of logical AND. Fixed #16
The logical expression syntax parser wasn't taking the logical AND
operator into account. As a result the cyclomatic complexity of "&&" was
miscalculated in the module.js unit tests.

This commit adds logic to increment the cyclomatic complexity given the
presence of logical ANDs as well as update the unit tests to reflect the
intended behavior.