Extended the top-level API to accept parsing overrides
Fixes #34

The top-level `analyse` function now accepts a third argument that can
be used to override the parsing behavior.

When given as a function it will be called for each code source and is
expected to return the resulting abstract syntax tree (AST). When the
given as an object, it will be used to override the options passed to
the default parser (esprima for the time being).

Removed the "perf" unit test for the project module. This is not a
replacement for a proper benchmark and is likely to increase development
friction.